pyproject.toml
[tool.poetry]
name = "backend-init"
version = "0.1.0"
description = ""
authors = ["SahanAtNeura <sahan@neuraintelligence.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.7"
uvicorn = "^0.25.0"
fastapi = "^0.105.0"
pydantic-settings = "^2.1.0"
beanie = "^1.23.6"


[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
isort = "^5.13.2"
flake8 = "^6.1.0"
mypy = "^1.7.1"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
mock = "^5.1.0"
pytest-asyncio = "^0.23.2"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
